Fix typo in formatting
